Integrate distributed/cifar10-100 into cifar10-100-collection.
