--- cifar10-100-collection/_checkpoint_nnp_util.py
@@ -1,25 +1,0 @@
-# Copyright 2020,2021 Sony Corporation.
-# Copyright 2021 Sony Group Corporation.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from __future__ import absolute_import
-
-import os
-import sys
-
-# Import save/load_checkpoint from utils
-sys.path.append(os.path.join(
-    os.path.dirname(__file__), '../utils/'))
-from neu.checkpoint_util import save_checkpoint, load_checkpoint
-from neu.save_nnp import save_nnp
