+++ cifar10-100-collection/checkpoint.py
