+++ cifar10-100-collection/cifar10_data.py
@@ -95,13 +95,13 @@ class Cifar10DataSource(DataSource):
 
     @property
     def images(self):
-        """Get copy of whole data with a shape of (N, 1, H, W)."""
-        return self._images.copy()
+        """Get whole data with a shape of (N, 3, H, W)."""
+        return self._images
 
     @property
     def labels(self):
-        """Get copy of whole label with a shape of (N, 1)."""
-        return self._labels.copy()
+        """Get whole label with a shape of (N, 1)."""
+        return self._labels
 
 
 def data_iterator_cifar10(batch_size,
@@ -116,8 +116,9 @@ def data_iterator_cifar10(batch_size,
     because :py:class:`Cifar10DataSource` is able to store all data into memory.
 
     '''
-    return data_iterator(Cifar10DataSource(train=train, shuffle=shuffle, rng=rng),
-                         batch_size,
-                         rng,
-                         with_memory_cache,
-                         with_file_cache)
+    data_source = Cifar10DataSource(train=train, shuffle=shuffle, rng=rng)
+    return data_source, data_iterator(data_source,
+                                      batch_size,
+                                      rng,
+                                      with_memory_cache,
+                                      with_file_cache)
