+++ cifar10-100-collection/hostfile
