--- distributed/cifar10-100/multi_device_multi_process_classification.sh
@@ -1,24 +1,0 @@
-# Copyright 2017,2018,2019,2020,2021 Sony Corporation.
-# Copyright 2021 Sony Group Corporation.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-#!/bin/bash
-
-# multi-gpu
-mpirun -n 4 python multi_device_multi_process_classification.py
-
-# multi-node
-#mpirun --hostfile hostfile python multi_device_multi_process_classification.py
-
-
