--- distributed/cifar10-100/requirements.txt
@@ -1,1 +1,0 @@
-nnabla>=0.9.9
